Fixes for repository

+++ ristoretoCoffeeShop/src/main/java/pl/sda/ristoretoCoffeeShop/model/Order.java
@@ -1,15 +1,12 @@
 package pl.sda.ristoretoCoffeeShop.model;
 
 import lombok.Data;
-import org.springframework.data.annotation.Id;
-import org.springframework.stereotype.Component;
+
+
 import pl.sda.ristoretoCoffeeShop.model.enums.OrderPaymentMethod;
 import pl.sda.ristoretoCoffeeShop.model.enums.OrderStatus;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Table;
+import javax.persistence.*;
 import java.util.Date;
 
 @Data
@@ -18,7 +15,7 @@ import java.util.Date;
 public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private OrderStatus status;
     private double total;
     private Date date;
